rtmp-services: Move service-specific files
Cleans up the directory a bit
