+++ plugins/rtmp-services/CMakeLists.txt
@@ -7,10 +7,10 @@ include_directories(${LIBCURL_INCLUDE_DIRS})
 include_directories(${OBS_JANSSON_INCLUDE_DIRS})
 
 set(rtmp-services_SOURCES
-	twitch.c
-	younow.c
-	nimotv.c
-	showroom.c
+	service-specific/twitch.c
+	service-specific/younow.c
+	service-specific/nimotv.c
+	service-specific/showroom.c
 	rtmp-common.c
 	rtmp-custom.c
 	rtmp-services-main.c)
@@ -23,10 +23,10 @@ if(WIN32)
 endif()
 
 set(rtmp-services_HEADERS
-	twitch.h
-	younow.h
-	nimotv.h
-	showroom.h
+	service-specific/twitch.h
+	service-specific/younow.h
+	service-specific/nimotv.h
+	service-specific/showroom.h
 	rtmp-format-ver.h)
 
 set(RTMP_SERVICES_URL
